backend/requirement.txt
flask
flask_cors
google.generativeai
python-dotenv